Use f16 activation storage in exported shaders

+++ src/models/exported_qwen3_asr/tensors/audio_encoder.py
@@ -328,7 +328,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='layer_norm',
-                spec=TensorSpec(dtype='float32', shape=(133, 896)),
+                spec=TensorSpec(dtype='float16', shape=(133, 896)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -341,7 +341,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='linear_1',
-                spec=TensorSpec(dtype='float32', shape=(133, 896)),
+                spec=TensorSpec(dtype='float16', shape=(133, 896)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -354,7 +354,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='reshape_2',
-                spec=TensorSpec(dtype='float32', shape=(133, 14, 64)),
+                spec=TensorSpec(dtype='float16', shape=(133, 14, 64)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -367,7 +367,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='linear_2',
-                spec=TensorSpec(dtype='float32', shape=(133, 896)),
+                spec=TensorSpec(dtype='float16', shape=(133, 896)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -380,7 +380,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='reshape_3',
-                spec=TensorSpec(dtype='float32', shape=(133, 14, 64)),
+                spec=TensorSpec(dtype='float16', shape=(133, 14, 64)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -393,7 +393,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='linear_3',
-                spec=TensorSpec(dtype='float32', shape=(133, 896)),
+                spec=TensorSpec(dtype='float16', shape=(133, 896)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -406,7 +406,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='reshape_4',
-                spec=TensorSpec(dtype='float32', shape=(133, 14, 64)),
+                spec=TensorSpec(dtype='float16', shape=(133, 14, 64)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -419,7 +419,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='transpose_1',
-                spec=TensorSpec(dtype='float32', shape=(14, 133, 64)),
+                spec=TensorSpec(dtype='float16', shape=(14, 133, 64)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -432,7 +432,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='unsqueeze',
-                spec=TensorSpec(dtype='float32', shape=(1, 14, 133, 64)),
+                spec=TensorSpec(dtype='float16', shape=(1, 14, 133, 64)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -445,7 +445,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='transpose_2',
-                spec=TensorSpec(dtype='float32', shape=(14, 133, 64)),
+                spec=TensorSpec(dtype='float16', shape=(14, 133, 64)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -458,7 +458,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='unsqueeze_1',
-                spec=TensorSpec(dtype='float32', shape=(1, 14, 133, 64)),
+                spec=TensorSpec(dtype='float16', shape=(1, 14, 133, 64)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -471,7 +471,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='transpose_3',
-                spec=TensorSpec(dtype='float32', shape=(14, 133, 64)),
+                spec=TensorSpec(dtype='float16', shape=(14, 133, 64)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -484,7 +484,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='unsqueeze_2',
-                spec=TensorSpec(dtype='float32', shape=(1, 14, 133, 64)),
+                spec=TensorSpec(dtype='float16', shape=(1, 14, 133, 64)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -497,7 +497,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='scaled_dot_product_attention',
-                spec=TensorSpec(dtype='float32', shape=(1, 14, 133, 64)),
+                spec=TensorSpec(dtype='float16', shape=(1, 14, 133, 64)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -510,7 +510,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='transpose_4',
-                spec=TensorSpec(dtype='float32', shape=(1, 133, 14, 64)),
+                spec=TensorSpec(dtype='float16', shape=(1, 133, 14, 64)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -523,7 +523,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='contiguous',
-                spec=TensorSpec(dtype='float32', shape=(1, 133, 14, 64)),
+                spec=TensorSpec(dtype='float16', shape=(1, 133, 14, 64)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -536,7 +536,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='reshape_5',
-                spec=TensorSpec(dtype='float32', shape=(133, 896)),
+                spec=TensorSpec(dtype='float16', shape=(133, 896)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -549,7 +549,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='linear_4',
-                spec=TensorSpec(dtype='float32', shape=(133, 896)),
+                spec=TensorSpec(dtype='float16', shape=(133, 896)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -562,7 +562,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='add_1',
-                spec=TensorSpec(dtype='float32', shape=(133, 896)),
+                spec=TensorSpec(dtype='float16', shape=(133, 896)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -575,7 +575,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='layer_norm_1',
-                spec=TensorSpec(dtype='float32', shape=(133, 896)),
+                spec=TensorSpec(dtype='float16', shape=(133, 896)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -588,7 +588,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='linear_5',
-                spec=TensorSpec(dtype='float32', shape=(133, 3584)),
+                spec=TensorSpec(dtype='float16', shape=(133, 3584)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -601,7 +601,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='gelu_3',
-                spec=TensorSpec(dtype='float32', shape=(133, 3584)),
+                spec=TensorSpec(dtype='float16', shape=(133, 3584)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -614,7 +614,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='linear_6',
-                spec=TensorSpec(dtype='float32', shape=(133, 896)),
+                spec=TensorSpec(dtype='float16', shape=(133, 896)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -627,7 +627,7 @@ def create_encoder_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='add_2',
-                spec=TensorSpec(dtype='float32', shape=(133, 896)),
+                spec=TensorSpec(dtype='float16', shape=(133, 896)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -895,7 +895,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key=None,
-                spec=TensorSpec(dtype='float32', shape=(11, 1, 128, 100)),
+                spec=TensorSpec(dtype='float16', shape=(11, 1, 128, 100)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.INPUT,
                 memory=MemoryClass.HOST_INPUT,
@@ -908,7 +908,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key=None,
-                spec=TensorSpec(dtype='float32', shape=(11, 13, 896)),
+                spec=TensorSpec(dtype='float16', shape=(11, 13, 896)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.INPUT,
                 memory=MemoryClass.HOST_INPUT,
@@ -934,7 +934,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key=None,
-                spec=TensorSpec(dtype='float32', shape=(1, 1, 133, 133)),
+                spec=TensorSpec(dtype='float16', shape=(1, 1, 133, 133)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.INPUT,
                 memory=MemoryClass.HOST_INPUT,
@@ -947,7 +947,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='conv2d',
-                spec=TensorSpec(dtype='float32', shape=(11, 480, 64, 50)),
+                spec=TensorSpec(dtype='float16', shape=(11, 480, 64, 50)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -960,7 +960,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='gelu',
-                spec=TensorSpec(dtype='float32', shape=(11, 480, 64, 50)),
+                spec=TensorSpec(dtype='float16', shape=(11, 480, 64, 50)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -973,7 +973,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='conv2d_1',
-                spec=TensorSpec(dtype='float32', shape=(11, 480, 32, 25)),
+                spec=TensorSpec(dtype='float16', shape=(11, 480, 32, 25)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -986,7 +986,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='gelu_1',
-                spec=TensorSpec(dtype='float32', shape=(11, 480, 32, 25)),
+                spec=TensorSpec(dtype='float16', shape=(11, 480, 32, 25)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -999,7 +999,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='conv2d_2',
-                spec=TensorSpec(dtype='float32', shape=(11, 480, 16, 13)),
+                spec=TensorSpec(dtype='float16', shape=(11, 480, 16, 13)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -1012,7 +1012,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='gelu_2',
-                spec=TensorSpec(dtype='float32', shape=(11, 480, 16, 13)),
+                spec=TensorSpec(dtype='float16', shape=(11, 480, 16, 13)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -1025,7 +1025,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='reshape',
-                spec=TensorSpec(dtype='float32', shape=(11, 7680, 13)),
+                spec=TensorSpec(dtype='float16', shape=(11, 7680, 13)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -1038,7 +1038,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='transpose',
-                spec=TensorSpec(dtype='float32', shape=(11, 13, 7680)),
+                spec=TensorSpec(dtype='float16', shape=(11, 13, 7680)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -1051,7 +1051,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='linear',
-                spec=TensorSpec(dtype='float32', shape=(11, 13, 896)),
+                spec=TensorSpec(dtype='float16', shape=(11, 13, 896)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -1064,7 +1064,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='add',
-                spec=TensorSpec(dtype='float32', shape=(11, 13, 896)),
+                spec=TensorSpec(dtype='float16', shape=(11, 13, 896)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -1077,7 +1077,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='reshape_1',
-                spec=TensorSpec(dtype='float32', shape=(143, 896)),
+                spec=TensorSpec(dtype='float16', shape=(143, 896)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -1090,7 +1090,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='index_select',
-                spec=TensorSpec(dtype='float32', shape=(133, 896)),
+                spec=TensorSpec(dtype='float16', shape=(133, 896)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -1103,7 +1103,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='layer_norm_36',
-                spec=TensorSpec(dtype='float32', shape=(133, 896)),
+                spec=TensorSpec(dtype='float16', shape=(133, 896)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -1116,7 +1116,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='linear_109',
-                spec=TensorSpec(dtype='float32', shape=(133, 896)),
+                spec=TensorSpec(dtype='float16', shape=(133, 896)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -1129,7 +1129,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='gelu_21',
-                spec=TensorSpec(dtype='float32', shape=(133, 896)),
+                spec=TensorSpec(dtype='float16', shape=(133, 896)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -1142,7 +1142,7 @@ def create_audio_encoder(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='linear_110',
-                spec=TensorSpec(dtype='float32', shape=(133, 1024)),
+                spec=TensorSpec(dtype='float16', shape=(133, 1024)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
