+++ src/models/exported_qwen3_asr/tensors/embed_tokens.py
@@ -71,7 +71,7 @@ def create_embed_tokens(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='embedding',
-                spec=TensorSpec(dtype='float32', shape=(1, 151, 1024)),
+                spec=TensorSpec(dtype='float16', shape=(1, 151, 1024)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
