+++ src/models/quantized_qwen3_asr/tensors/audio_inject.py
@@ -62,7 +62,7 @@ def create_audio_inject(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key=None,
-                spec=TensorSpec(dtype='float32', shape=(audio_sequence_length, 1024)),
+                spec=TensorSpec(dtype='float16', shape=(audio_sequence_length, 1024)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.INPUT,
                 memory=MemoryClass.HOST_INPUT,
@@ -75,7 +75,7 @@ def create_audio_inject(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='unsqueeze',
-                spec=TensorSpec(dtype='float32', shape=(1, audio_sequence_length, 1024)),
+                spec=TensorSpec(dtype='float16', shape=(1, audio_sequence_length, 1024)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -88,7 +88,7 @@ def create_audio_inject(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='index_copy',
-                spec=TensorSpec(dtype='float32', shape=(1, sequence_length, 1024)),
+                spec=TensorSpec(dtype='float16', shape=(1, sequence_length, 1024)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
